requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
django-crispy-forms>=1.4.0,<3.0
Pillow>=6.0,<7.0
django-tastypie-hmacauth==0.0.1b5
